% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autor_trial.R
\name{autor_trial}
\alias{autor_trial}
\title{Title}
\usage{
autor_trial(
  trial,
  family,
  dispersion = 1,
  size = 1,
  de = 1,
  ie = 0,
  range = 0.9,
  sqrt_n_fixed = 20,
  neighbor_dist = 1
)
}
\arguments{
\item{trial}{x}

\item{family}{x}

\item{dispersion}{x}

\item{size}{x}

\item{n_fixed}{x}
}
\value{
x
}
\description{
Title
}
