% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpred.R
\name{estpred}
\alias{estpred}
\title{prediction and estimation after estimating covariance parameters}
\usage{
estpred(
  theta,
  y,
  sampsize = NULL,
  X,
  distmat = NULL,
  autocor_fun,
  maxvar,
  maxrange,
  maxphi = NULL,
  family,
  Z1 = NULL,
  Z2 = NULL,
  mask = NULL,
  Xp = NULL,
  Z1p = NULL,
  Z2p = NULL,
  dist_op = NULL,
  dist_pp = NULL,
  maskop = NULL,
  maskpp = NULL,
  W = NULL,
  M = NULL,
  use.nugget = TRUE
)
}
\arguments{
\item{theta:}{covariance parameters in order 1) nugget, 2) partial sill, 3) range these are on logit scale, and are converted to max*expit(theta\link{i}) internally where max is maxvar or maxrange}

\item{y:}{vector of response variable}

\item{X:}{design matrix of explanatory variables}

\item{distmat:}{matrix of Euclidean distance among all points}

\item{autocor_fun:}{geostatistical autocorrelation function that take distmat and range parameter as it arguments}

\item{maxvar:}{maximum variance for either partial sill or nugget effect}

\item{maxrange:}{maximum for the range parameter}

\item{family:}{family for the data model, one of either "poisson", "binomial", "negbinomial", "gamma", or "invgaussian"
Xp: design matrix of explanatory variables for prediction locations}

\item{dist_op:}{Euclidean distance between observed and prediction locations}

\item{dist_pp:}{Euclidean distance among prediction locations}
}
\description{
prediction and estimation after estimating covariance parameters
}
\author{
Jay Ver Hoef
}
