% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat_trial.R
\name{geostat_trial}
\alias{geostat_trial}
\title{Title}
\usage{
geostat_trial(
  trial,
  family,
  dispersion = 1,
  size = 1,
  n_fixed = 300,
  sp_bayes = FALSE
)
}
\arguments{
\item{trial}{x}

\item{family}{x}

\item{dispersion}{x}

\item{size}{x}

\item{n_fixed}{x}
}
\value{
x
}
\description{
Title
}
