% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_from_list.R
\name{get_autor_output_from_list}
\alias{get_autor_output_from_list}
\title{Title}
\usage{
get_autor_output_from_list(object, only_time = FALSE)
}
\arguments{
\item{object}{x}
}
\value{
x
}
\description{
Title
}
