% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhglmm.laplace-package.R
\docType{package}
\name{mhglmm.laplace-package}
\alias{mhglmm.laplace}
\alias{mhglmm.laplace-package}
\title{mhglmm.laplace: Repository for Marginal Inference for Hierarchical Generalized Linear Mixed Models with Patterned Covariance Matrices Using the Laplace Approximation by Jay M. Ver Hoef, Eryn Blagg, Michael Dumelle, Philip Dixon, Dale L. Zimmerman, and Paul Conn.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Authors:
\itemize{
  \item Jay M. Ver Hoef \email{jay.verhoef@noaa.gov} (\href{https://orcid.org/0000-0003-4302-6895}{ORCID})
}

}
\keyword{internal}
