% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_trial.R
\name{verify_trial}
\alias{verify_trial}
\title{Title}
\usage{
verify_trial(
  trial,
  family,
  dispersion = 1,
  size = 1,
  n_fixed = 300,
  nugget = FALSE
)
}
\arguments{
\item{trial}{x}

\item{family}{x}

\item{dispersion}{x}

\item{size}{x}
}
\value{
x
}
\description{
Title
}
